[Feat] 뷰 컴포넌트화 및 코드 분리
